requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Integrations
openai==1.3.0
anthropic==0.7.0
requests==2.31.0
aiohttp==3.9.0

# Search and Web Scraping
google-search-results==2.4.2
beautifulsoup4==4.12.2
trafilatura==1.6.0
requests-cache==1.1.1

# CLI and UI
typer==0.9.0
rich==13.7.0
prompt-toolkit==3.0.43

# Web UI (FastAPI)
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sse-starlette==1.8.2
markdown==3.5.1
pymdown-extensions==10.5
pygments==2.17.2
aiosqlite==0.19.0
slowapi==0.1.9

# Async Support
asyncio-contextmanager==1.0.1

# Code Execution Safety
# restricted-execution

# Utilities
pyyaml==6.0.1
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-benchmark==4.0.0
pytest-timeout==2.2.0
psutil==5.9.6

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.0
autoflake==2.3.1
vulture==2.14

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==1.25.2
myst-parser==2.0.0

# RAG Framework (Phase 1)

# Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1

# Embeddings & Reranking (Phase 2)
transformers==4.37.2
torch==2.2.0

# Document Processing
pypdf==3.17.4
pymupdf==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0

# Domain-Specific Tools (Phase 3)
pyowm==3.3.0                 # Weather API
yfinance==0.2.35             # Finance (fallback)
alpha-vantage==2.3.1         # Finance (primary)
openrouteservice==2.3.3      # Routing/Transportation

# Multimodal Support (Phase 4)
paddleocr==2.10.0             # OCR for Chinese/English
paddlepaddle==3.2.1          # PaddleOCR backend (CPU)
google-generativeai==0.8.5   # Gemini vision API
pillow==12.0.0               # Image processing

# Optional: Reranking API
# cohere==4.47               # Faster reranking (paid)
